qcacld-3.0: Move sta state to not connected after try disconnect
qcacld-2.0 to qcacld-3.0 propagation

If a connect is received with previous connect already in progress,
sta connect state is moved from connecting to disconnecting and
disconnect is initiated to stop the previous connect.

If previous is present in sme/roam pending queue the disconnect
will abort the previous connect by removing the cmd from pending
queue. In this case the connect state is not moved back to
not connected from disconecting.

Thus the current connect command fails and all new connect command
will fail as the connect state will always be in disconnecting state.

To avoid this, move the connect state to not connected, once try
disconnect is issued internally.

Change-Id: Ia563f85df84cb9dbb3d3a82bc97127d39e7b1421
CRs-Fixed: 961559