qcacld-3.0: Do not request ROC if already on correct channel
This is qcacld-2.0 to qcacld-3.0 propagation

If cfg80211_mgmt_tx_params.offchan is set, driver need not send ROC
if device is already on correct channel. Check current home channel
of associated adapter and if it is same as target channel requested
skip ROC request and directly send frame over the air.

Change-Id: Ib13a885437a4349702e1706b3a466af262813f8a
CRs-Fixed: 921117