qcacld-3.0: Fix Ext Cap IE mismatch in Probe/Assoc Response frames
qcacld-2.0 to qcacld-3.0 propagation

In SAP mode, while sending Probe response template to HAL from driver,
ext cap IE should be merged only if it is present or received from hostapd.
In function limSendProbeRspTemplateToHal, extcap_present bool is wrongly
initialized to true which leads to updation of the extended Cap IE with
junk values in Probe response even for the cases where driver is not
supposed to update it, i.e. cases where the ext cap IE is not present or
not received from hostapd

Git-commit: a7256f3483c0a26c8715f2ae6166289de6da6b1e
Change-Id: If82fbba4c20b5e96dbe1b11c97ab557ded488f3e
CRs-Fixed: 995545