qcacld-3.0: Fix potential reference of uninitialized roam profile
Currently in hdd_send_re_assoc_event() there are several guard tests
which protect the function from executing unless all of the
prerequisite information is present. If any of the tests fail then the
code performs centralized cleanup. One of the cleanups performed is to
free the roam profile.  Unfortunately at the time most of the guard
tests are performed the roam profile has not been retrieved from SME,
and has not been initialized in any other way.  This would ultimately
result in csr_roam_free_connect_profile() reading "garbage" values for
pBssDesc and pAddIEAssoc.  Address this issue by relocating the
existing qdf_mem_zero() of the roam profile so that the roam profile
is initialized before any of the guard tests are performed.

Change-Id: I2a2ba92ffb5b73c3ba4f54482d8f2b37f6e2dad6
CRs-Fixed: 1076422