qcacld-3.0: Check the device status during hdd cleanup of present mode
On a change of the mode ( from the current one to the FTM mode ) ,
the clean up of the adapters are again done though they are already
cleaned ( signified by device_status = DRIVER_MODULES_CLOSED ). Hence
check for this device_status while to avoid a redundant clean up of the
adapters.

Change-Id: I2a851b235809db5cd47ec5d738067d1553f83129
CRs-Fixed: 1115243