qcacld-3.0: Clear the HDD scan list during exit
HDD scan list is not cleared as part of
wlan driver exit if the driver state is
closed.

Clear the HDD scan list irrespective of the driver
status but send scan abort to firmware only when
firmware is active. It mean don't send abort when
driver is in closed state.

Change-Id: Ib1b52203aafafd3e7af50224e5d245c80c45efad
CRs-Fixed: 2137288