asoc: sdm845: check sound card status before sending AFE config
WCD codec must be enumerated before AFE configuration for
some of the features can be sent to DSP. This is because
as part of setting up this configuration, DSP tries to
access some codec register. These register accesses will
fail if codec is not enumerated already.
Add a check for Sound Card online state to confirm that WCD
codec has already enumerated before sending AFE configs.

Change-Id: I0b7e5eb49e4f6c6aee9cb142054e57eae592602b
Signed-off-by: Banajit Goswami <bgoswami@codeaurora.org>