qcacld-3.0: Start SAP on 80Mhz in STA+SAP when acs_policy enabled
When acs_policy is enabled and STA is already up on some 5Ghz channel,
SAP is coming up on the same 5Ghz channel as expected but in 20Mhz even
when bandwidth given by hostapd is 80Mhz.

Start SAP on same channel as STA but in 80Mhz instead of 20Mhz.

Change-Id: Ie5dc06844f867d32796706cdcc190d1cb1b1d2fe
CRs-Fixed: 2196229