qcacld-3.0: Add wifi config vtable support
One of the HDD functions with the highest cyclomatic complexity is
__wlan_hdd_cfg80211_wifi_configuration_set(). In order to reduce the
complexity there is a plan to replace the inline attribute handling
with a vtable-based approach

As the first step of this refactoring add the actual vtable
infrastructure.  Later changes will refactor the attribute handling
and populate the vtables.

Change-Id: If79fc61f84bf17a39e300c18681961a4a0e01408
CRs-Fixed: 2371558