qcacld-3.0: Harmonize lim_process_add_sta_rsp()
The parameter identifiers used in the implementation of
lim_process_add_sta_rsp() differ from the ones used in the prototype.
Since this can be confusing for both humans and code analysis tools,
harmonize on the parameters used in the implementation. In addition,
relocate the documentation so that it is the interface that is
documented and not the implementation.

Change-Id: I221b4fd0a643770b46adf06776b16fa85a6de03c
CRs-Fixed: 2376110