qcacld-3.0: Add a missing increment for fw module id
When host driver is setting the fw log level, there is a chance
for the loop to be unable to exit when a certain fw module's
log level is incorrectly set.

Add an additional increment to the fw module id when a certain
fw module's log level is incorrectly set.

Change-Id: Id5a0268536a4d6b74732515dac43f583174c9e51
CRs-Fixed: 2378208