qcacmn: Increase fw_txrx rem stats counter size
rem_stats keeps counter of remaining stats to be processed.
At present this counter is of type uint8_t and hence can
accumulate till 256 counter but num_stats is of type uint32_t.
Hence if remainting stats to be precessed is 256 then rem_stats
value will be 0 as it will only consider LSB 8 BITS and it will
result in stats mismatch and hence memory leak.

Increasing size to uint64_t to accumulate more counter size.

Change-Id: Ifa86c7e4a73a5a700e9033c8756e986d86025cf9