msm: camera: icp: Fix AHB, AXI voting in icp
Due to a mismatched internal structure num_paths for
AXI voting is always set to 0 for ipe, bps which is
causing not to vote any AXI bandwidth for these hws.
After all contexts associated with a device go to
inactive state, reduce the AHB vote to minimum value
to make sure previously voted high AHB vote doesn't
remain for the rest of camera session.

CRs-Fixed: 2546088
Change-Id: I7621bb85280b2541ed37a87c9e486dd6be62ea24
Signed-off-by: Pavan Kumar Chilamkurthi <pchilamk@codeaurora.org>