qcacld-3.0: Return the number of bytes placed in to the buffer
When the sysfs file 'gtx_bw_mask' is read, DEVICE_ATTR 'show' callback
function is called and currently its returning 'ret' which is 0 and hence
the content of sysfs file "/sys/class/net/wlanxx/gtx_bw_mask" is empty.

Hence the fix is 'show' callback function should return the number of
bytes placed in to the buffer.

Change-Id: I44b7eda40cd2963b4be52f97b513badc48244120
CRs-Fixed: 2720141