rmnet: change to correctly account for non-linear IPv6 SKBs
Packets received from physical device in rmnet_data layer
can be non-linear. The existing logic for handling packets
in low-latency data path works only for linear SKBs.
This fix ensures that both linear and non-linear SKBs are
handled and header pointer can be obtained correctly so
that the packet can be sent up the network stack for further
processing.

Signed-off-by: Kaustubh Pandey <quic_kapandey@quicinc.com>
Change-Id: Icef6235a0b1141fa3defd8d883069fde2a336c12