qca-wifi: Deprecate the ext_ifu_acs configuration
Deprecate the per-VAP ext_ifu_acs configuration. This is required in
order to prevent internal and external ACS from running in parallel if
the user mis-configures some VAPs for internal ACS and some for
external ACS. Instead, an INI for global configuration is provided in
a different, related change.

Change-Id: I714e7a65d7e1ebd3f73479b51ffddc474af7f74c
CRs-Fixed: 2757150