qcacld-3.0: fix CMEM de-initialization
DP SOC's fst_in_cmem will be set to true when the FW has support
for placing FST in CMEM; in FTM mode, though this variable will be
set, CMEM parameters will not be initialized. When the mode is
changed to mission mode, we try to de-initialize CMEM parameters
which are not initialized. Do not rely on this variable for
de-initialization of CMEM parameters.

Change-Id: I1818d5572b281b4ae800f3fef0c633fedde9e758
CRs-Fixed: 2783939