qcacld-3.0: Do RTPM get in SWLM flush timer
In runtime suspended state, tx packets cause a
runtime resume and if packets are received in
a way that tput pass count exceeds the threshold
but bytes flush and time threshold are not
reached, the flush timer is started. As part of
dp_swlm_tcl_flush_timer, iowrite can happen even
before the runtime resume is completed resulting
in a NOC error.

Fix is to do runtime_get in dp_swlm_tcl_flush_timer
before doing iowrite.

Change-Id: I83b6d8cfc7b29bc1f30fda94007b3663d6a99405
CRs-Fixed: 2783826