Revert "disp: msm: sde: fix race between disable commit and vblank work"
This is a partial revert of commit:
	86f79244e3800f8c7373a3ec7c7e013d36ac2c86.

The enc->crtc link cannot be relied upon outside of the display
thread context since the DRM legacy modeset API will detach and
re-attach this connection on every commit.

The logic to flush the event thread in all cases during the CRTC
disable is kept from the original patch.

Change-Id: I2cd4ae47589bb397c72de0e9deba9b0c710d379a
Signed-off-by: Steve Cohen <cohens@codeaurora.org>