msm: camera: isp: Choose appropriate WM for frame header
Frame header is not enabled on all WMs. Currently frame header
is piggy backed with the first WM for which buffer is configured.
It is possible that this WM does not support frame header.
This change tries with all WMs till it finds a WM that supports
frame header.

CRs-Fixed: 2823154
Change-Id: I300cb449cdeebc74300d100dfdd8dbe50bec9bce
Signed-off-by: Karthik Anantha Ram <kartanan@codeaurora.org>