qcacld-3.0: Don't Check for concurrency if pre_cac adapter exists
Currently, host driver is checking for concurrency while handling
the pre_cac request. pre_cac adapter won't be deleted until SAP is
stopped. As a result the second pre cac request for the same SAP
will be dropped due to this concurrency check.

To avoid this, Don't check for pre cac concurrency if the pre cac
adapter is already exists.

Change-Id: I780c6310cf819da1a2f5c7c93590ccb1b544e7a4
CRs-Fixed: 2848736