qca-wifi: Move dfs_set_nol back to common code
The dfs_set_nol feature is moved to WIN code as part of an FR58190.
But the dfs_set_nol feature is used by MCL. Therefore moving the
dfs_set_nol feature back to common code so that both MCL and Host can
use the dfs_set_nol feature.

CRs-Fixed: 2883810
Change-Id: I4bfa55cbf984f612043a11893704603d3e282cb5