qcacld-3.0: Enable single-target SPF support
Currently the SPF support is triggered when there is more than one
target specified in TARGET_WLAN_CHIP. For taro we know we need
multi-target support in the future, but we only want to support a
single target during bringup. However we want all procedures
established during the bringup process to be applicable to the final
product. Therefore update the logic such that SPF support is triggered
whenever the TARGET_WLAN_CHIP is anything other than the default value
of 'wlan'.

Change-Id: I0a0671d2ebbdb85985bc57376b4acd2bd79a16ae
CRs-Fixed: 2890488