msm: camera: isp: Change last_dump_err_req field to a bool
In case of streams with no ePCR, and if error happens on the
very first frame reg dump on error is not triggered since
last_dump_err_req_id and applied_req_id both are 0x0. Change
it to a flag instead to avoid this scenario and reset the flag
at stream off/release.

CRs-Fixed: 2841729
Change-Id: Ieaee68c42fb2228dfa966a0ef158802a10694dca
Signed-off-by: Karthik Anantha Ram <kartanan@codeaurora.org>