qcacld-3.0: Add retries to state control file creation
wlan_hdd_state_ctrl_param_create() creates a control file that is used
to synchronize the wlan driver with the Wi-Fi HAL in userspace. When
multiple instances of the driver are loaded in parallel, only one can
create and own the state ctrl param. The first instance of the driver
that is loaded creates the state ctrl param, and then it will wait for
a certain amount of time to be probed. If it is probed, then that
instance of the driver will stay loaded and no other instances of the
driver can load. But if it is not probed, then that instance of the
driver will destroy the state ctrl param and exit, and another
instance of the driver can then create the state ctrl param.

In the current code if a second instance of the driver tries to create
the state ctrl param while the first instance is still waiting for
the probe, then the second instance will fail and will exit. If the
first instance subsequently isn't probed and times out, then no driver
will be loaded.

To address this issue add retries to the state control file creation
so that the second instance will not exit after one failure, but
instead exits only after so much time has passed that the first
instance of the driver must have been probed.

Change-Id: I35cf04145d610613a51e5321603d48d988c16479
CRs-Fixed: 2905637