disp: msm: sde: handle empty list of ROI in rounded corner
Currently, RC is assuming that partial update callback functions
will not be called when there is a full frame update and that CP
framework will return early instead. This assumption had to be changed
to support full frame transition for CP features such as demura and SPR
and now the responsibility is up to each feature to handle this corner
case.

This change is returning early in PU callback function if the payload
buffer is 0.

Change-Id: I478888fb036e642d5274a2cccc4d5378e62d6afa
Signed-off-by: Amine Najahi <anajahi@codeaurora.org>