qcacld-3.0: Update exteneded capabilities after connection
The extended capabilities IE is sent during vdev create and
this doesn't have the merged extended capabilities from the
userspace. So some of the fields advertised in association
request are not advertised in the re-association request.

Fix the missing fields not set in the extended capabilities IE
by sending the set IE message to LIM after connection is
successful.

Change-Id: I322d057f602e0362f12b362e7818e84cbb056cf4
CRs-Fixed: 2971072