qcacmn: Set "agile_precac_active" for all applicable pdevs
Currently, "agile_precac_active" is set only after receiving the start
event in the init state. "agile_precac_active" indicates whether adfs is
supported in the radio or not. This has to be set based on the agile precac
enabled check when the vap is brought up and when the event
DFS_AGILE_SM_EV_AGILE_START is sent to DFS Agile State Machine.

Change-Id: I2084d6d413ee11fa9f77026326dab79aafcd64fb