qcacmn: Set extended supported rates IE max len to 12
As per the spec, the maximum IE length of supported rates are 8 and ext
supported rates are max_rates minus supported rates (=44-8).
In the field, some APs are non-compliance with the spec and includes
12 rates. Due to this, our STAs does not scan and connect to those APs.

To fix this issue, supported rates IE length is already set to 8. Hence,
set extended supported rates IE max len back to 12.

Change-Id: I0d8ce972b08aa28c08a7e797df4208dd36154983
CRs-Fixed: 3036878