msm: camera: isp: Handle RUP for enabled CSID CAMIF IRQs
CAMIF interrupts in CSID are enabled based on use case requirements.
In additionto this, inside hw manager propagation to isp context is
again dependent on the same checks. Since the enabling is based
on use-case, the checks in the handler functions are redundant.
With upgrades in hardware, the checks are increasing while handling
in hw manager.
This commit removes the checks in handler function in hw manager.

CRs-Fixed: 3052221
Change-Id: I44f59b782dba0f7e7eb20c2ace56f3094527932b
Signed-off-by: Gaurav Jindal <gjindal@codeaurora.org>