qcacld-3.0: Fix connectivity logging roam result and roam reason
Roam result is printed as failure always since the
is_roam_successful flag is not passed to the
roam successful attribute to userspace. For roam
reason, the wrong attribute is used to convert the roam
sub reason from driver values to userspace values.
In WTC roam, the sub reason & vsie is not filled
from the right roam trigger TLV.

Fix the connectivity logging roam result and roam
subreason and WTC values passed to userspace.

Change-Id: Id276054868d907a801e87c1fd9cd7bb3b7c4e150
CRs-Fixed: 3051447