qcacld-3.0: Fix EAP length attribute type
EAP length attribute is of type NLA_U8, but the table entry
has this attribute type as NLA_U16. This causes parsing
failure of this attribute and the length print is not
included in the logs.

Fix the NLA type for the QCA_WLAN_VENDOR_ATTR_DIAG_EAP_LEN
attribute in the connectivity logging table.

Change-Id: I3570931799811787169c9403b1f26488379524bc
CRs-Fixed: 3054835