msm: ipa: Add API to set IPA client filtering table SRAM priority high
IPA client filtering table can reside in SRAM in case there is enough
space. Since SRAM is a limited resource it might also reside in slower
memory (DDR). This new API enables setting an IPA client's filtering
table SRAM priority to high. It also includes an IOCTL that enables
to use the API from user-space.
Moving an IPA client filtering table to SRAM greatly
improves access time to the filtering rules in it and therefore
the entire data path of this client.
Current limitation is this API needs to be called before filter
rules are installed.

Change-Id: I34814369b9c4f6ea535b739aed9a20df8606080b
Acked-by: Eliad Ben Yishay <ebenyish@qti.qualcomm.com>
Signed-off-by: Ilia Lin <quic_ilial@quicinc.com>