qcacld-3.0: Fix EAP length and roam candidate info
EAP length is of 2 bytes, but currently the attribute type
used is NLA_U8. This causes the EAP packets with length
greater than 255 to print invalid length.
For roam current AP, the roamed AP info is printed.

Use NLA_U16 attribute for the EAP length and fill
correct event type for the current connected AP.

Change-Id: I9b86055a4817163db17e469ca45b4a0044d5b2d6
CRs-Fixed: 3075012