qcacld-3.0: Increase the max supported sta config to 4 for HMT
Currently host does not send the max supported the sta configuration
for all targets, it is defaulted to the 3 in the firmware.
Firmware treats all the P2P DEVICE mode as sta internally so when
MLO is enabled with a legacy STA and P2P device mode the firmware
is rejecting because max sta  configuration is reached.

To support this ML_STA+STA+P2P increase the max supported the sta
configuration to 4 in the WMI_INIT_CMDID.

Change-Id: I699c370200eab673d3e035f7f6b41e3e88202401
CRs-Fixed: 3081160