qcacld-3.0: Reject LL stats request for SAP mode
LL stats request is not supported for SAP mode. But driver is not checking
for device mode for LL stats request. Instead it is depending on LL stats
request set flag which won't be set for SAP mode. In case of mode switch
from STA to SAP, it is possible that LL stats request set flag will be set
for STA mode and LL stats are requested after switching to SAP mode. This
will bypass the mode SAP.

To avoid this, reject LL stats request for SAP mode.

Change-Id: I8c680288991126f0b349584f502b22e18af72287
CRs-Fixed: 3081453