qcacld-3.0: Reduce lock held time
Currently we process and free the nbuf from
queue while holding a lock for the complete duration.
There might be a case when there are many packets pending
in the queue. This might cause the issue when we held the lock
and other process waits for lock.

So as part of the fix only perform the nbuf unlink inside the
lock and move the processing and free part outside the lock.

Change-Id: Iea0cabecc25ce90c189454b288421b1e720777f2
CRs-Fixed: 3125790