qcacld-3.0: Don't disconnect if not in connected state
When the country changes, there is a check to see if the current phy
mode or freq or ch width is not allowed, in which case, disconnect
the adapter. However, this check can run even if the adapter is not
currently connected. Add an extra check to see if it is even connected
first, before checking if disconnection is necessary.

Change-Id: Ic4243fd85205e930e49d6f8f9f0e4399851e870a
CRs-Fixed: 3139370