disp: msm: sde: disable CWB crop params as part of wb_disable
WB crop params are supported only in CWB mode and not for
normal writeback. So these crop params should be disabled
at the end of a CWB session. If this is not done, WB
timeouts can happen if a normal WB commit comes after a
CWB crop commit. This change adds support to disable the
crop params during wb disable whenever clone mode is
enabled.

Change-Id: Icb3792d13728ac25d232198eb59f0cc737dfbbe4
Signed-off-by: Krishna Manikandan <quic_mkrishn@quicinc.com>