msm: camera: memmgr: Avoid TOCTOU buffer access during multiple use of same fd
Fd is a user-accessible value, referring it multiple times leads to
TOCTOU issues. Dma_buf can be freed after the 1st use of fd and create
another different dma_buf but with same fd. In such scenario, during 2nd
use of fd, we will get a different dma_buf with different len. To avoid
this, we can use same dma_buf instead of retrieving it twice using same fd.
In this change FD is accessed only once in syscall.

CRs-Fixed: 3159446
Change-Id: I00eb6dd3d798165f5c6c0bd59feabe80a68592b1
Signed-off-by: Yash Upadhyay <quic_yupadhya@quicinc.com>