qcacld-3.0: Start SAP on valid channel after country change
Currently if SAP is up on 5 GHz DFS/non-DFS channel and country
change happens to world mode where SAP is not allowed on 5 GHz
channels then it moves to a 2.4 GHz channel and saves current
operating frequency band information, now again if a country
change happens SAP tries to move back to the first 5 GHz
active/passive channel.

Sap should come to active channel first. If not available then
it should come on DFS channel, but it shouldn't come up on
passive channel.

To address this issue, add a fix to filter out passive channels.

Change-Id: I32e8a81c2a5d9e785262edbeeef2c879c87f1302
CRs-Fixed: 3277965