dp: Support single GPIO aux switch
Before this change, only the "flip" GPIO is used to decide whether to
turn on/off the GPIO aux switch logic, and it is assumed that both the
"flip" and "enable" GPIOs would be used at the same time.

This change makes it possible to turn on the aux switch logic when only
one GPIO is used.

Change-Id: I42b6d1ea1eaa176f7e0737197a2ca2d110da3b39
Signed-off-by: Qing Chang <qing@motorola.com>
Reviewed-on: https://gerrit.mot.com/2557645
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Konstantin Makariev <kmakariev@motorola.com>
Submit-Approved: Jira Key