qcacld-3.0: Use reg_is_chan_enum_invalid API to check for invalid channels
Currently the INVALID_CHANNEL assumes the value 0xBAD. When a channel list
array is indexed, there is an array out of bound error when the channel
list array is indexed using INVALID_CHANNEL or when the index is greater
than equal to NUM_CHANNELS. So a check for an invalid channel enum is
likely to be, (ch_enum == INVALID_CHANNEL) && (ch_enum  >= NUM_CHANNEL).
To address this issue, use reg_is_chan_enum_invalid API to check if
the channel is invalid.

Change-Id: Ie1fa97242607deff08d6987b955d2d88b534b44c
CRs-Fixed: 3271922