qcacld-3.0: Properly handle the flags defined as modules
The current way in the driver to handle CONFIG_CNSS_UTILS=m and
CONFIG_WCNSS_MEM_PRE_ALLOC=m will not work if the flags are not
coming from kernel Kconfig directly (e.g. CNSS utils and prealloc
module sources are built out of kernel tree) since "cppflags-
$(CONFIG_FLAG) += -DCONFIG_FLAG" will not take effect if CONFIG_FLAG=m.
Fix it in a proper way to work even if config is defined as m.

Change-Id: I60c26b1e0d64277b733e534960a00afb5579fb43
CRs-Fixed: 3419566