video: driver: update buffer count considering prior updated value
Buffer count is defaulted to a certain codec specific value with the
assumption that it is first time the port is streamed on. For cases,
when a port is streamed on after stream_off, firmware might have
already updated the desired buffer count. In such scenario, retaining
the default count would lead to mismatch in driver and firmware. For
ex. the COMV size might be based on certain number of minimum buffers,
and then driver calculates with default count. This would result in
mismatch when driver sets the size to firmware when streamed on.
It is desired to consider the maximum of default count and the count
while the port was streaming before streamed off.

Change-Id: I3f5b8750f92064548af508aa883365f8d78dbbf2
Signed-off-by: Vikash Garodia <quic_vgarodia@quicinc.com>